[WORDPRESS-41] Removing duplicated lines between components by fusioning them
