+++ react/components/WordpressAllPosts.tsx
@@ -1,41 +1,14 @@
 /* eslint-disable @typescript-eslint/camelcase */
-import { Container } from 'vtex.store-components'
-import React, {
-  ChangeEvent,
-  Fragment,
-  useState,
-  useEffect,
-  useRef,
-} from 'react'
-import { Helmet } from 'react-helmet'
-import { defineMessages, useIntl } from 'react-intl'
-import { useQuery } from 'react-apollo'
-import { useRuntime } from 'vtex.render-runtime'
-import { Spinner, Pagination } from 'vtex.styleguide'
-import { useCssHandles } from 'vtex.css-handles'
+import React from 'react'
+import { defineMessages } from 'react-intl'
 
-import WordpressTeaser from './WordpressTeaser'
-import Settings from '../graphql/Settings.graphql'
-import AllPosts from '../graphql/AllPosts.graphql'
-import WordpressCategorySelect from './WordpressCategorySelect'
-import WordpressTagSelect from './WordpressTagSelect'
-
-const CSS_HANDLES = [
-  'listTitle',
-  'listContainer',
-  'listFlex',
-  'listFlexItem',
-  'paginationComponent',
-  'filtersContainer',
-  'categorySelectContainer',
-  'tagSelectContainer',
-] as const
+import WordpressPosts from './WordpressPosts'
 
 interface AllPostsProps {
   customDomain: string
   customDomainSlug: string
   subcategoryUrls: boolean
-  mediaSize?: MediaSize
+  mediaSize: MediaSize
   postsPerPage: number
 }
 
@@ -46,284 +19,15 @@ const WordpressAllPosts: StorefrontFunctionComponent<AllPostsProps> = ({
   mediaSize,
   postsPerPage,
 }) => {
-  const intl = useIntl()
-  const {
-    route: { id, params },
-    pages,
-    query,
-    setQuery,
-    navigate,
-  } = useRuntime() as any
-  const initialPage = params.page ?? query?.page ?? '1'
-  const [page, setPage] = useState(parseInt(initialPage, 10))
-  const [perPage, setPerPage] = useState(postsPerPage)
-  const [selectedOption, setSelectedOption] = useState(postsPerPage)
-  const [selectedCategory, setSelectedCategory] = useState('')
-  const [categories, setCategories] = useState([])
-  const [selectedTag, setSelectedTag] = useState('')
-  const [tags, setTags] = useState([])
-  const handles = useCssHandles(CSS_HANDLES)
-  const { loading: loadingS, data: dataS } = useQuery(Settings)
-  const { loading, error, data, fetchMore } = useQuery(AllPosts, {
-    variables: {
-      wp_page: 1,
-      wp_per_page: perPage,
-      customDomain,
-    },
-  })
-
-  const containerRef = useRef<null | HTMLElement>(null)
-  const initialPageLoad = useRef(true)
-  const [loadingPage, setLoadingPage] = useState(true)
-
-  useEffect(() => {
-    setLoadingPage(true)
-  }, [page])
-  useEffect(() => {
-    data && setLoadingPage(false)
-  }, [data])
-
-  useEffect(() => {
-    if (initialPageLoad.current) {
-      initialPageLoad.current = false
-
-      return
-    }
-
-    if (containerRef.current) {
-      window.scrollTo({
-        top:
-          containerRef.current.getBoundingClientRect().top +
-          window.pageYOffset -
-          100,
-        behavior: 'smooth',
-      })
-    }
-  }, [page])
-  useEffect(() => {
-    data &&
-      setCategories(
-        data.wpPosts.posts
-          .reduce((acc: any, el: any) => [...acc, ...el.categories], [])
-          .reduce((acc: any, current: any) => {
-            const x = acc.find((item: any) => item.id === current.id)
-            return !x ? acc.concat([current]) : acc
-          }, [])
-      )
-    data &&
-      setTags(
-        data.wpPosts.posts
-          .reduce((acc: any, el: any) => [...acc, ...el.tags], [])
-          .reduce((acc: any, el: any) => {
-            const x = acc.find((item: any) => item.id === el.id)
-            return !x ? acc.concat([el]) : acc
-          }, [])
-      )
-  }, [data])
-
-  const PaginationComponent = (
-    <Pagination
-      rowsOptions={[
-        postsPerPage,
-        postsPerPage * 2,
-        postsPerPage * 3,
-        postsPerPage * 4,
-      ]}
-      selectedOption={selectedOption}
-      currentItemFrom={(page - 1) * perPage + 1}
-      currentItemTo={page * perPage}
-      textOf="of"
-      textShowRows={
-        dataS?.appSettings?.displayShowRowsText === false
-          ? null
-          : // eslint-disable-next-line @typescript-eslint/no-use-before-define
-            intl.formatMessage(messages.postsPerPage)
-      }
-      totalItems={data?.wpPosts?.total_count ?? 0}
-      onRowsChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-        setPage(1)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = '1'
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: '1' })
-        }
-        setSelectedOption(+value)
-        setPerPage(+value)
-        fetchMore({
-          variables: {
-            wp_page: 1,
-            wp_per_page: +value,
-            customDomain,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-      onPrevClick={() => {
-        if (page <= 1) return
-        const prevPage = page - 1
-        setPage(prevPage)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = prevPage.toString()
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: prevPage.toString() })
-        }
-        fetchMore({
-          variables: {
-            wp_page: prevPage,
-            wp_per_page: perPage,
-            customDomain,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-      onNextClick={() => {
-        const nextPage = page + 1
-        setPage(nextPage)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = nextPage.toString()
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: nextPage.toString() })
-        }
-        fetchMore({
-          variables: {
-            wp_page: nextPage,
-            wp_per_page: perPage,
-            customDomain,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-    />
-  )
-
   return (
-    <Container
-      className={`${handles.listContainer} pt6 pb8`}
-      style={{ maxWidth: '90%' }}
-      ref={containerRef}
-    >
-      {dataS?.appSettings?.titleTag && (
-        <Helmet>
-          <title>{dataS.appSettings.titleTag}</title>
-        </Helmet>
-      )}
-      <div className={`${handles.paginationComponent} ph3`}>
-        {PaginationComponent}
-      </div>
-      <div
-        className={`${handles.filtersContainer} flex flex-row justify-between mt3 ph3`}
-      >
-        {dataS?.appSettings?.filterByCategories && (
-          <div className={`${handles.categorySelectContainer} w-40`}>
-            <WordpressCategorySelect
-              categories={categories}
-              selectedCategory={selectedCategory}
-              setSelectedCategory={setSelectedCategory}
-            />
-          </div>
-        )}
-        {dataS?.appSettings?.filterByTags && (
-          <div className={`${handles.tagSelectContainer} w-40`}>
-            <WordpressTagSelect
-              selectedTag={selectedTag}
-              setSelectedTag={setSelectedTag}
-              tags={tags}
-            />
-          </div>
-        )}
-      </div>
-      {(loading || loadingS || loadingPage) && (
-        <div className="mv5 flex justify-center" style={{ minHeight: 800 }}>
-          <Spinner />
-        </div>
-      )}
-      {error && (
-        <div className="ph5" style={{ minHeight: 800 }}>
-          Error: {error.message}
-        </div>
-      )}
-      {data?.wpPosts ? (
-        <Fragment>
-          <div className={`${handles.listFlex} mv4 flex flex-row flex-wrap`}>
-            {data.wpPosts.posts
-              .filter(
-                (post: any) =>
-                  (selectedCategory && selectedCategory !== 'all'
-                    ? post.categories.find(
-                        (category: any) => category.name === selectedCategory
-                      )
-                    : post) &&
-                  (selectedTag && selectedTag !== 'all'
-                    ? post.tags.find(
-                        (tag: any) => tag.id === parseInt(selectedTag, 10)
-                      )
-                    : post)
-              )
-              .map((post: PostData, index: number) => (
-                <div
-                  key={index}
-                  className={`${handles.listFlexItem} mv3 w-100-s w-50-l ph4`}
-                >
-                  <WordpressTeaser
-                    title={post.title.rendered}
-                    author={post.author ? post.author.name : ''}
-                    categories={post.categories}
-                    subcategoryUrls={subcategoryUrls}
-                    excerpt={post.excerpt.rendered}
-                    date={post.date}
-                    id={post.id}
-                    slug={post.slug}
-                    link={post.link}
-                    customDomainSlug={customDomainSlug}
-                    featuredMedia={post.featured_media}
-                    mediaSize={mediaSize}
-                    showAuthor={false}
-                    showCategory
-                    showDate
-                    showExcerpt
-                    useTextOverlay={false}
-                    absoluteLinks={false}
-                  />
-                </div>
-              ))}
-          </div>
-          <div className={`${handles.paginationComponent} ph3 mb7`}>
-            {PaginationComponent}
-          </div>
-        </Fragment>
-      ) : (
-        !loading &&
-        !error && (
-          <div>
-            <h2>No posts found.</h2>
-          </div>
-        )
-      )}
-    </Container>
+    <WordpressPosts
+      customDomain={customDomain}
+      customDomainSlug={customDomainSlug}
+      subcategoryUrls={subcategoryUrls}
+      mediaSize={mediaSize}
+      postsPerPage={postsPerPage}
+      categoryVariable={null}
+    />
   )
 }
