+++ react/components/WordpressCategory.tsx
@@ -1,20 +1,15 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import { Container } from 'vtex.store-components'
-import React, {
-  ChangeEvent,
-  Fragment,
-  useState,
-  useEffect,
-  useRef,
-} from 'react'
+import React, { Fragment, useState, useEffect, useRef } from 'react'
 import { useQuery } from 'react-apollo'
-import { defineMessages, useIntl } from 'react-intl'
+import { defineMessages } from 'react-intl'
 import { useRuntime } from 'vtex.render-runtime'
-import { Spinner, Pagination } from 'vtex.styleguide'
+import { Spinner } from 'vtex.styleguide'
 import Helmet from 'react-helmet'
 import { useCssHandles } from 'vtex.css-handles'
 
 import WordpressTeaser from './WordpressTeaser'
+import PaginationComponent from './PaginationComponent'
 import CategoryPostsBySlug from '../graphql/CategoryPostsBySlug.graphql'
 import Settings from '../graphql/Settings.graphql'
 
@@ -37,13 +32,10 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
   postsPerPage,
   mediaSize,
 }) => {
-  const intl = useIntl()
   const {
-    route: { id, params },
-    pages,
+    route: { params },
+
     query,
-    setQuery,
-    navigate,
   } = useRuntime() as any
 
   let parsedCustomDomains = null
@@ -62,7 +54,6 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
   const initialPage = params.page ?? query?.page ?? '1'
   const [page, setPage] = useState(parseInt(initialPage, 10))
   const [perPage, setPerPage] = useState(postsPerPage)
-  const [selectedOption, setSelectedOption] = useState(postsPerPage)
   const categoryVariable = {
     categorySlug:
       params.subcategoryslug_id ||
@@ -109,107 +100,6 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
     }
   }, [page])
 
-  const PaginationComponent = (
-    <Pagination
-      rowsOptions={[
-        postsPerPage,
-        postsPerPage * 2,
-        postsPerPage * 3,
-        postsPerPage * 4,
-      ]}
-      selectedOption={selectedOption}
-      currentItemFrom={(page - 1) * perPage + 1}
-      currentItemTo={page * perPage}
-      textOf="of"
-      textShowRows={
-        dataS?.appSettings?.displayShowRowsText === false
-          ? null
-          : // eslint-disable-next-line @typescript-eslint/no-use-before-define
-            intl.formatMessage(messages.postsPerPage)
-      }
-      totalItems={data?.wpCategories?.categories[0]?.wpPosts?.total_count ?? 0}
-      onRowsChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-        setPage(1)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = '1'
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: '1' })
-        }
-        setSelectedOption(+value)
-        setPerPage(+value)
-        fetchMore({
-          variables: {
-            wp_page: 1,
-            wp_per_page: +value,
-            customDomain,
-            ...categoryVariable,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-      onPrevClick={() => {
-        if (page <= 1) return
-        const prevPage = page - 1
-        setPage(prevPage)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = prevPage.toString()
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: prevPage.toString() })
-        }
-        fetchMore({
-          variables: {
-            wp_page: prevPage,
-            wp_per_page: perPage,
-            customDomain,
-            ...categoryVariable,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-      onNextClick={() => {
-        const nextPage = page + 1
-        setPage(nextPage)
-        if (pages[id].path.indexOf(':page') > 0) {
-          params.page = nextPage.toString()
-          navigate({
-            page: id,
-            params,
-            scrollOptions: false,
-          })
-        } else {
-          setQuery({ page: nextPage.toString() })
-        }
-        fetchMore({
-          variables: {
-            wp_page: nextPage,
-            wp_per_page: perPage,
-            customDomain,
-            ...categoryVariable,
-          },
-          updateQuery: (prev, { fetchMoreResult }) => {
-            if (!fetchMoreResult) return prev
-            return fetchMoreResult
-          },
-        })
-      }}
-    />
-  )
   return (
     <Fragment>
       {dataS && data?.wpCategories?.categories?.length > 0 && (
@@ -232,7 +122,18 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
         ref={containerRef}
       >
         <div className={`${handles.paginationComponent} ph3`}>
-          {PaginationComponent}
+          <PaginationComponent
+            postsPerPage={postsPerPage}
+            page={page}
+            perPage={perPage}
+            dataS={dataS}
+            data={data}
+            setPage={setPage}
+            setPerPage={setPerPage}
+            fetchMore={fetchMore}
+            customDomain={customDomain}
+            categoryVariable={categoryVariable}
+          />
         </div>
         {(loading || loadingS || loadingPage) && (
           <div className="mv5 flex justify-center" style={{ minHeight: 800 }}>
@@ -276,7 +177,18 @@ const WordpressCategory: StorefrontFunctionComponent<CategoryProps> = ({
               )}
             </div>
             <div className={`${handles.paginationComponent} ph3 mb7`}>
-              {PaginationComponent}
+              <PaginationComponent
+                postsPerPage={postsPerPage}
+                page={page}
+                perPage={perPage}
+                dataS={dataS}
+                data={data}
+                setPage={setPage}
+                setPerPage={setPerPage}
+                fetchMore={fetchMore}
+                customDomain={customDomain}
+                categoryVariable={categoryVariable}
+              />
             </div>
           </Fragment>
         ) : (
