+++ react/components/withSearchContext.tsx
@@ -1,5 +1,4 @@
 import { SearchPageContext } from 'vtex.search-page-context/SearchPageContext'
-
 import React, { ComponentType } from 'react'
 
 function withSearchContext(WrappedComponent: ComponentType<any>) {
